Some updates.

+++ src/commands/rm.ts
@@ -25,7 +25,14 @@ interface RmFlags {
 async function executeRemove(cwd: string, args: string[]) {
   const flags = parseArgs(args);
   await Promise.all(flags.paths.map((specifiedPath) => {
+    if (specifiedPath.length === 0) {
+      throw new Error("Bug in dax. Specified path should have not been empty.");
+    }
     const path = resolvePath(cwd, specifiedPath);
+    if (path === "/") {
+      // just in case...
+      throw new Error("Cannot delete root directory. Maybe bug in dax? Please report this.");
+    }
     return Deno.remove(path, { recursive: flags.recursive });
   }));
 }
@@ -40,33 +47,33 @@ export function parseArgs(args: string[]) {
 
   for (const arg of parse_arg_kinds(args)) {
     if (
-      (arg.arg == "recursive" && arg.kind == "LongFlag")
-      || (arg.arg == "r" && arg.kind == "ShortFlag")
-      || (arg.arg == "R" && arg.kind == "ShortFlag")
+      (arg.arg === "recursive" && arg.kind === "LongFlag")
+      || (arg.arg === "r" && arg.kind == "ShortFlag")
+      || (arg.arg === "R" && arg.kind === "ShortFlag")
     ) {
       result.recursive = true;
     } else if (
-      (arg.arg == "dir" && arg.kind == "LongFlag")
-      || (arg.arg == "d" && arg.kind == "ShortFlag")
+      (arg.arg == "dir" && arg.kind === "LongFlag")
+      || (arg.arg == "d" && arg.kind === "ShortFlag")
     ) {
       result.dir = true;
     } else if (
-      (arg.arg == "force" && arg.kind == "LongFlag")
-      || (arg.arg == "f" && arg.kind == "ShortFlag")
+      (arg.arg == "force" && arg.kind === "LongFlag")
+      || (arg.arg == "f" && arg.kind === "ShortFlag")
     ) {
       result.force = true;
     } else {
-      if (arg.kind != "Arg") bailUnsupported(arg);
-      result.paths.push(arg.arg);
+      if (arg.kind !== "Arg") bailUnsupported(arg);
+      result.paths.push(arg.arg.trim());
     }
   }
-  if (result.paths.length == 0) {
+  if (result.paths.length === 0) {
     throw Error("missing operand");
   }
   return result;
 }
 
-function bailUnsupported(arg: ArgKind) {
+function bailUnsupported(arg: ArgKind): never {
   switch (arg.kind) {
     case "Arg":
       throw Error(`unsupported argument: ${arg.arg}`);
